requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.17.0
alembic==1.11.1
altair==5.2.0
annotated-types==0.5.0
anyio==3.7.0
appnope==0.1.3
APScheduler==3.10.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-lru==2.0.4
async-timeout==4.0.2
asyncpg==0.27.0
asynctest==0.13.0
atlassian-python-api==3.40.0
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
blinker==1.7.0
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
comm==0.1.3
confection==0.0.4
coverage==7.2.3
cssselect==1.2.0
cymem==2.0.7
dataclasses-json==0.5.8
datasets==2.14.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.7
discord==2.3.1
discord-py-interactions==5.8.0
discord-py-slash-command==4.2.1
discord-typings==0.6.0
discord.py==2.3.1
distlib==0.3.6
distro==1.8.0
dnspython==2.3.0
docker==6.0.1
docstring-parser==0.15
docutils==0.20.1
dotty-dict==1.3.1
duckduckgo-search==2.8.6
ecdsa==0.18.0
email-validator==2.0.0.post2
emoji==2.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
environs==9.5.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.100.1
fastapi-admin==1.0.4
fastjsonschema==2.18.0
feedparser==6.0.10
filelock==3.12.2
flake8==6.0.0
flupy==1.2.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-search-results==2.4.2
googleapis-common-protos==1.60.0
gotrue==1.0.2
greenlet==2.0.2
grpcio==1.53.0
gTTS==2.3.1
h11==0.14.0
html2text==2020.1.16
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.16.4
identify==2.5.22
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
invoke==1.7.3
ipykernel==6.25.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
iso8601==1.1.0
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keyring==24.2.0
langchain==0.0.248
langchainplus-sdk==0.0.9
langcodes==3.3.0
langsmith==0.0.16
llama-hub==0.0.19
llama-index @ git+https://github.com/Moi1oM/llama_index_supabase.git@d62ff29510a171a22a251718b1a331550a4f7c0a
loguru==0.7.0
lxml==4.9.2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
midjourney-api==1.0.7
mistune==3.0.1
more-itertools==10.0.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
nodeenv==1.7.0
notebook==7.0.1
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==1.1.1
openai-function-calling==1.1.1
openapi-schema-pydantic==1.2.4
orjson==3.9.0
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathspec==0.11.1
pathy==0.10.1
pendulum==2.1.2
pexpect==4.8.0
pgvector==0.1.8
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
pkginfo==1.9.6
platformdirs==3.10.0
playsound==1.2.2
pluggy==1.0.0
postgrest==0.10.6
pre-commit==3.2.2
preshed==3.0.8
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.2
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.4.0
pydeck==0.8.1b0
pyflakes==3.0.1
Pygments==2.15.1
pymilvus==2.2.11
PyMySQL==1.0.3
pyparsing==3.1.1
pypika-tortoise==0.1.6
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.15.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-semantic-release==7.33.2
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
readability-lxml==0.8.1
readme-renderer==40.0
realtime==1.0.0
redis==4.5.4
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
scipy==1.11.1
selenium==4.9.0
semver==2.13.0
Send2Trash==1.8.2
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqladmin==0.13.0
SQLAlchemy==2.0.16
srsly==2.4.6
sse-starlette==1.6.1
stack-data==0.6.2
starlette==0.27.0
storage3==0.5.3
streamlit==1.30.0
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
tenacity==8.2.2
termcolor==2.3.0
terminado==0.17.1
thinc==8.1.9
tiktoken==0.4.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.1
tornado==6.3.2
tortoise-orm==0.19.3
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.13.0
twine==3.8.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
ujson==5.7.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
validators==0.22.0
vecs==0.3.0
virtualenv==20.22.0
wasabi==0.10.1
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.6.1
websockets==10.4
widgetsnbextension==4.0.8
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.0.1
xxhash==3.3.0
yarl==1.9.2
youtube-transcript-api==0.5.0
zipp==3.16.2
